Don't redeploy the test API server on every CircleCI run (#331)

+++ api/libproject/devstart.rb
@@ -730,6 +730,19 @@ def with_cloud_proxy_and_db_env(cmd_name, args)
 end
 
 def circle_deploy(cmd_name, args)
+  # See https://circleci.com/docs/1.0/environment-variables/#build-details
+  common = Common.new
+  common.status "circle_deploy with branch='#{ENV.fetch("CIRCLE_BRANCH", "")}'" +
+  " and tag='#{ENV.fetch("CIRCLE_TAG", "")}'"
+  if ENV.has_key?("CIRCLE_BRANCH") and ENV.has_key?("CIRCLE_TAG")
+    raise("expected exactly one of CIRCLE_BRANCH and CIRCLE_TAG env vars to be set")
+  end
+  is_master = ENV.fetch("CIRCLE_BRANCH", "") == "master"
+  if !is_master and !ENV.has_key?("CIRCLE_TAG")
+    common.status "not master or a git tag, nothing to deploy"
+    return
+  end
+
   unless Workbench::in_docker?
     exec *(%W{docker run --rm -v #{File.expand_path("..")}:/w -w /w/api
       allofustest/workbench:buildimage-0.0.9
@@ -742,7 +755,7 @@ def circle_deploy(cmd_name, args)
     exit 1
   end
 
-  if ENV["CIRCLE_BRANCH"] == "master"
+  if is_master
     common.status "Running database migrations..."
     with_cloud_proxy_and_db_env(cmd_name, args) do
       migrate_database
@@ -751,8 +764,21 @@ def circle_deploy(cmd_name, args)
     end
   end
 
-  version = ENV.fetch("CIRCLE_TAG", "circle-ci-test")
-  promote = ENV["CIRCLE_BRANCH"] == "master" ? "--promote" : "--no-promote"
+  promote = ""
+  version = ""
+  if is_master
+    # Note that --promote will generally be a no-op, as we expect
+    # circle-ci-test to always be serving 100% traffic. Pushing to an existing
+    # live version will immediately make those changes live. In the event that
+    # someone mistakenly pushes a different version manually, this --promote
+    # will restore us to the expected circle-ci-test version on the next commit.
+    promote = "--promote"
+    version = "circle-ci-test"
+  elsif ENV.has_key?("CIRCLE_TAG")
+    promote = "--no-promote"
+    version = ENV["CIRCLE_TAG"]
+  end
+
   deploy(cmd_name, args + %W{--version #{version} #{promote}})
 end
